requirements.txt
absl_py==2.1.0
datasets==2.20.0
einops==0.8.0
hydra-core==1.3.2
immutabledict==4.2.0
langdetect==1.0.9
nltk==3.8.1
numpy==2.0.0
omegaconf==2.3.0
pandas==2.2.2
plotly==5.22.0
torch==2.3.1
tqdm==4.66.4
transformer_lens==2.2.2
transformers==4.42.3
